% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffa_screen_plot 20260421.R
\name{ch_ffa_screen_plot}
\alias{ch_ffa_screen_plot}
\title{ch_ffa_screen_plot}
\usage{
ch_ffa_screen_plot(
  df,
  mtitle = NULL,
  stn,
  polar = TRUE,
  kthresh = 5,
  thresh = NULL,
  n = 12,
  m = 0,
  M = 2,
  offset = 0,
  trans = "log",
  mcol = c("orange", "gray60", "red", "darkred")
)
}
\arguments{
\item{df}{dataframe with date of event (maxdate) and annual maximum (amax)}

\item{mtitle}{title for plot}

\item{stn}{stationID}

\item{polar}{if TRUE [default] a polar plot of the annual maxima and outliers is produced.}

\item{kthresh}{kappa threshold default is 5}

\item{thresh}{POT threshold default is NULL}

\item{n}{default is 12, number of colours for number of months}

\item{m}{smallest angle in radians parameter for generating circular colours default = 0}

\item{M}{largest angle in radians parameter for generating circular colours default = 2 (*pi)}

\item{offset}{the zero in radians, default is 0.}

\item{trans}{transform series peak. options are "none", "log", and "boxcox"}

\item{mcol}{array of three colours, default is c("orange", "gray70", "red") use to emphasize
low, normal, and high values}
}
\value{
a list containing
\describe{
  \item (station) (stationID)
  \item (nyears) (number of years of data)
  \item (maxQ) (maximum flow - Flood of Record)
  \item (mindex) (year with maximum flow)
  \item (maxDate) (Date of Flood of Record)
  \item (maxdoy) (day of year with FoR)
  \item (WalfWolfowitz) (result of WW test against trend)
  \item (MannKendall) (result of Mann Kendall trend test)
  \item (Pettitt) (result of Pettitt test for single change point)
  \item (POTmodel) (model of POT)
  \item (out_hgrubb) (number of high outliers, indexes in df)
  \item (out_lgrubb) (number of low outliers, indexes in df)
  \item (out_timing) (number of timing outliers, indexes in df)
  \item (fz_out) (indexes of fuzzy clusters outliers, indexes in df)
  \item (df) (a dataframe with results for each year)
 }
Also produces an annotated three plots: an annotated time series plot,
an return period plot of flood events and an optional polar plot
}
\description{
generate a flood frequency plot with symbols indicating whether an observation is a
high or low outlier, and colour symbols indicating month of the year in which the flood
occurred.
}
\examples{
\donttest{
# Not tested automatically as can be very slow to execute
data(CAN05AA008)
amax <- ch_get_amax(CAN05AA008)
ch_ffa_screen_plot(amax)}
}
\references{
Cohn, T. A., J. F. England, C. E. Berenbrock, R. R. Mason, J. R.
Stedinger and J. R. Lamontagne (2013). "A generalized Grubbs‐Beck test statistic
for detecting multiple potentially influential low outliers in flood series."
Water Resources Research 49(8): 5047-5058 10.1002/wrcr.20392: 10.1002/wrcr.20392.

Sau, M. F. and D. Rodriguez (2018). "Minimum distance method for directional data and outlier detection." Advances in Data Analysis and Classification 12: 587-603 DOI: 10.1007/s11634-017-0287-9.

Whitfield, P. H. and D. H. Burn (2025*). "Extreme and Rogue Floods in North America."
Journal of Hydrology. in review.
}
