% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_ECDE_metadata_v2.R
\name{ch_get_ECDE_metadata_v2}
\alias{ch_get_ECDE_metadata_v2}
\title{new version of ch_ECDE_metadata}
\usage{
ch_get_ECDE_metadata_v2(filename, writefile = NULL)
}
\description{
new version of ch_ECDE_metadata
}
