% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_plot_POT.R
\name{ch_plot_POT}
\alias{ch_plot_POT}
\title{ch_plot_POT}
\usage{
ch_plot_POT(
  object,
  npy,
  main,
  xlab,
  ylab,
  xlimsup,
  mon,
  ccol,
  mcex,
  mpch,
  mcode,
  sr_pch,
  mcol,
  isout,
  ci = TRUE,
  points = TRUE,
  ...
)
}
\arguments{
\item{object}{output from POT function fitgpd}

\item{npy}{number of events per year = #events/#years}

\item{main}{plot title}

\item{xlab}{xaxis label}

\item{ylab}{y axis label}

\item{xlimsup}{upper limit for Pareto}

\item{mon}{array of months of observations [1-12] +}

\item{ccol}{array of circular colours [12] +}

\item{mcex}{size of plot symbols [4] +}

\item{mpch}{symbols for points [4] +}

\item{mcode}{array of indexes for symbol plotting [1-4] +}

\item{sr_pch}{symbols for overlay [NA, 8] +}

\item{mcol}{array of colours for outline of points [3] +}

\item{isout}{array of indexes for timing outliers [1-2] +}

\item{ci}{confidence interval = TRUE}

\item{points}{plot points = TRUE}

\item{...}{}
}
\value{
returns the pot function invisibly and generates a plot
}
\description{
A plot of peaks-over-threshold derived from a plot method from the package POT.

A single object matching ‘retlev.uvpot’ was found. It was found in the following places
registered S3 method for retlev from namespace POT
namespace:POT
with value

The original function only plotted generic symbols for the observations.
It was modified to add informative symbols and colours for FFA screening. These
are indicated with a "+" below.
The plots are shown in Whitfield and Burn (2025b)
}
\examples{
\dontrun{}
\donttest{}

}
\references{
Whitfield, P. H. and D. H. Burn (2025*). "Screening Annual Maxima and
Peaks-Over-Threshold Series for Flood Frequency Analysis."
}
\author{
modified by Paul Whitfield original from POT
}
