% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_hstat.R
\name{SR_hstat}
\alias{SR_hstat}
\title{SR_hstat}
\usage{
SR_hstat(x, alpha = 0.05)
}
\arguments{
\item{x}{An array of positions on the unit circle in radians}

\item{alpha}{a confidence limit.}
}
\value{
a list containing
\itemize{
  \item (mu) (circular number of mean directions)
  \item (ventana) (circular number of window)
  \item (muini) (circular number initial value of mu )
  \item (conc) (estimate of the concentration parameter defined in Ko (1992).)
  \item (malpha) (1 - alpha)
  \item (distance) (array of distance to be compared to corte)
  \item (corte) (cut value)
  \item (isout) (sequential indexes of outlier wherein 0 is an outlier and a 1 is not)
  \item (porcout) (proportion of outliers in sample)
 }
}
\description{
@description These are an updated version of the functions described in
 Sau & Rodrigues (2018).  Daniela Rodriguez kindly shared their original code.
 The main modifications here were that the data tables are not read in, but included
 as variables and most of the language was changed to English from Spanish.
}
\details{
This function is basically a wrapper for the series of functions written
 by Sau & Rogrigues, and these revisions were made by Paul Whitfield.
 Originally separate functions, they were nested with their parent.

 Therefore there are three main functions and the execution code within.
 All the original comments were preserved subsequent to conversion
 to English letter equivalents; these are in comment lines starting with ##.
}
\examples{
data1 <- circular::rwrappedcauchy(100, mu=circular(0), rho=0.7, control.circular=list(units="degrees"))
result <- SR_hstat(data1, alpha)
result$isout # lists the outliers with 0 being an outlier and 1 not.
}
\references{
Sau, M.F., and D. Rodriguez. 2018. "Minimum distance method for
directional data and outlier detection."  Advances in Data Analysis
and Classification 12:587-603.

Ko D (1992) Robust estimation of the concentration parameter of the von Mises-Fisher
distribution. Ann Stat 20:917–928

Whitfield, P. H. and D. H. Burn (2026 in review). "Technical Note: A Visual
Diagnostic Framework for Identifying Non-Stationarity and Mixed Populations in Flood Series"
}
\author{
Mercedes Fernandez Sau, Daniela Rodrigues, Paul Whitfield
}
